+++ pypideploy.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+rm -rf dist/*
 python setup.py sdist
 python setup.py bdist_wheel --universal
 twine upload dist/*